Have killed off the impact of fp1, fp2, fp4 on DECAY widths and then added it globally to each partial width *after* the BSM contributions, according to the higgs squared rescaling
1 / (1 + fp2 * vev^2)

This has also fixed an inconsistency in the  h>aa and h>az, where the
higgs renormalization was in the fermion couplings in the loop, but
not in the boson couplings. This previously led to an increase of the
h>aa and h>az widths when fp2 was switched on due to the reduced destructive
interference from the fermions, rather than the expected decrease.

M    HDim6/inputs.cc
M    HDim6/inputs.h
M    HDim6/decay.cc