Add Stop of Fittino if an error occurs in Checkmate
Only possible now to figure out if an error occured, not yet which error. In addition the Calculator now uses two Inputfiles, one in which you just specify which analyses CheckMate shoul use and one in which you set-up the whole set-up. A few more changes have been made concerning which numbers from CheckMate are saved in the Output RootFiles.